+++ src/GRA.Controllers/Base/Controller.cs
@@ -16,6 +16,7 @@ using GRA.Abstract;
 
 namespace GRA.Controllers.Base
 {
+    [ResponseCache(Location = ResponseCacheLocation.None, NoStore = true)]
     [ServiceFilter(typeof(SiteFilter), Order = 1)]
     [SessionTimeoutFilter]
     public abstract class Controller : Microsoft.AspNetCore.Mvc.Controller