+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Remove username when users deleted
 - Remove secret code when trigger deleted
 - Change trigger search to include secret code
+- Disable page caching
 
 ### Fixed
 - Add field to site object to allow forcing https even if the Web server believes the request came in via http.